Use float value

+++ binkapp/Screens/Transactions/View/TransactionView.swift
@@ -22,7 +22,7 @@ class TransactionView: CustomView {
     }
     
     func configure(with transaction: CD_MembershipTransaction) {
-        let transactionValue = transaction.formattedAmounts?.first?.value?.intValue ?? 0
+        let transactionValue = transaction.formattedAmounts?.first?.value?.floatValue ?? 0
         let timestamp = transaction.timestamp?.doubleValue ?? 0.0
         let timestampDate = Date(timeIntervalSince1970: timestamp)
         if let transactionDescription = transaction.transactionDescription {
@@ -40,7 +40,7 @@ class TransactionView: CustomView {
         setValueLabel(text: "\(prefix)%.02f", transactionValue: transactionValue, addDecimals: true)
     }
       
-    func setValueLabel(text: String, transactionValue: Int, addDecimals: Bool) {
+    func setValueLabel(text: String, transactionValue: Float, addDecimals: Bool) {
         if transactionValue < 0 {
             let value = abs(transactionValue)
             valueLabel.text = "-" + String(format: text, addDecimals ? Float(value) : value)
